Split sensor.py into three focused modules
sensor.py was 1411 lines mixing URL decoding, markdown formatting, and
sensor entity logic. Extracted into:

- decoding.py  (106 lines): URL decode helpers for homework/weekplan data
- formatting.py (664 lines): markdown builders, date helpers, subject
  aliases, schedule formatting, derived homework helpers (pure functions)
- sensor.py    (618 lines): ConfigEntry-dependent helpers + all sensor
  entity classes + async_setup_entry

No logic changed. All 75 tests still pass.
